requirements.txt
Django==1.9
djangorestframework==3.5.3
Markdown==2.6.7
